Add Vue Carousel dependency.
Vue Carousel is used for cycling thru pictures. It is necessary for
displaying pictures of Imprimis gameplay using what seems to be the same
image element.

Adding this dependency was necessary because running `yarn dev` returned
a dependency warning, and the website also showed a warning. Instead of
a nice image carousel, visitors would get a nasty warning message.