Automatically pusblish webpage to gh-pages.
This commit generates a static project and uploads it to gh-pages.

In this way GitHub can host the static project without needing an
independent server. This saves both money and time.

Note that if the workflow does not have permissions to read and write,
it will not be able to overwrite the gh-pages branch and the website
will not be updated.