Specify base URL for static project.
Hosting the static project in GitHub did not work because images and
scripts did not load. The base URL when using gh-pages is the same as
the repo name. Changing the base URL to match the name of the
repo will hopefully fix broken URLs so that images and scripts load
properly.