Add language selector component.
A language selector allows visitors to view the page in one of the
supported languages.

The webpage had a Polish translation that was unrearchable because there
was no language selector. This commit fixes that problem by adding a
menu for selecting supported translations.

The language switcher uses Nuxt and i18n to iterate and show available
locales. This means adding new locales does not require modifying vue,
html, or css files.

Note that the language switcher is a component, meaning that it makes the
code reausable and cleaner. The language switcher can be added to any
page with a single line of code.